move census related test to census module
